Introduce Dittosuite repo
Dittosuite is a benchmarking and workload replication tool.
Although the project is already available in AOSP, adding this to the
kernel aims at reducing the gap between the tool and the kernel itself,
and reduce the development distance of both the projects.
Dittosuite will be created in //test/dittosuite, as it is already in
AOSP.

Test: repo sync
Bug: 322744630
Change-Id: Iabf0e61a0aae3279d5805097729f27e6ebf6343b
Signed-off-by: Alessio Balsini <balsini@google.com>