Mark external deps as ddk-external.
Mark selected external dependencies as ddk-external
group instead of the ddk group. This consists of
projects that are registered on BCR, so technically,
they are not needed unless for truly local builds.

The mirrored BCR is also marked as ddk-external because
a DDK module can use --registry=http://registry.bazel.build
to override the flag from bzlmod.bazelrc.

Drop stardoc because it is a dev_dependency of @kleaf
and usually isn't needed.

Add rules_rust for consistency with android15-6.6
even though rust has not been enabled on mainline.

Test: TH
Bug: 338439996
Change-Id: I78a1e53fc97e79352e87ee5d58d2676e7322d5ac